Reorganization of the scanner implementation and documentation
