+++ Scanners/DTL Scanner/.classpath
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="lib" path="/Users/dennisbehm/_temp/dbb_20_libraries/dbb.core_2.0.0.jar"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
